src/main/webapp/app/entities/command/route/command-routing-resolve.service.ts
import { Injectable } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { Resolve, ActivatedRouteSnapshot, Router } from '@angular/router';
import { Observable, of, EMPTY } from 'rxjs';
import { mergeMap } from 'rxjs/operators';

import { ICommand, Command } from '../command.model';
import { CommandService } from '../service/command.service';

@Injectable({ providedIn: 'root' })
export class CommandRoutingResolveService implements Resolve<ICommand> {
  constructor(protected service: CommandService, protected router: Router) {}

  resolve(route: ActivatedRouteSnapshot): Observable<ICommand> | Observable<never> {
    const id = route.params['id'];
    if (id) {
      return this.service.find(id).pipe(
        mergeMap((command: HttpResponse<Command>) => {
          if (command.body) {
            return of(command.body);
          } else {
            this.router.navigate(['404']);
            return EMPTY;
          }
        })
      );
    }
    return of(new Command());
  }
}
